Getting Line recognition and selecting. Still need to translate and rotate

+++ src/cs355/model/shape/Circle.java
@@ -1,21 +1,20 @@
 package cs355.model.shape;
 
 import java.awt.Color;
-import java.awt.Point;
 
 public class Circle extends Shape{
 	
-	private int radius;	
+	private double radius;	
 	
 	public Circle(Color color) {
 		super(color);
 	}
 
-	public int getRadius() {
+	public double getRadius() {
 		return radius;
 	}
 
-	public void setRadius(int radius) {
+	public void setRadius(double radius) {
 		this.radius = radius;
 	}
