+++ src/cs355/model/shape/Ellipse.java
@@ -5,22 +5,22 @@ import java.awt.Point;
 
 public class Ellipse extends Shape{
 	
-	private int height;
-	private int width;
+	private double height;
+	private double width;
 	
 	public Ellipse(Color color) {
 		super(color);
 	}
-	public int getHeight() {
+	public double getHeight() {
 		return height;
 	}
-	public void setHeight(int height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
-	public int getWidth() {
+	public double getWidth() {
 		return width;
 	}
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
 }
